+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added functionality to disable/enable window-managment for minimal-tiling. Features are now disabled dy default and require user to enable via the extensions settings panel.
 - Modified the defaults for some of shortcuts, including: (1) New-empty-workspace shortcut and (2) Window-mangement-assitant shortcuts
 ## New
-- Added ability to change shortcuts via `dconf` app and `gsettings` command
+- Added basic functionality to modify shortcuts via the setting's panel
 
 # 46.1 (Mar 2024)
 ## CHANGES